feat: bulk resume import with S3 storage and duplicate detection
- Add bulk upload endpoint for candidates (/api/candidates/bulk-upload)
- Add S3 storage service with local fallback for resume files
- Replace AI-based resume parsing with fast algorithmic extraction
  - Regex-based email, phone, name extraction
  - Keyword matching for skills
  - Date range analysis for experience years
- Add duplicate detection using SHA-256 file hash
- Add multi-file and folder selection to FileUpload component
- Add Candidates tab to BulkImport page (now default)
- Fix screening_analysis score type (int -> float)

+++ server/app/models/interview.py
@@ -60,7 +60,7 @@ class ConversationAnalysis(BaseModel):
 
 # Screening Analysis Models
 class ScreeningAttribute(BaseModel):
-    score: int  # 0-100
+    score: float  # 0-100
     evidence: list[str]
     notes: Optional[str] = None
 
@@ -70,7 +70,7 @@ class ScreeningAnalysis(BaseModel):
     engagement_energy: ScreeningAttribute
     critical_thinking: ScreeningAttribute
     professionalism: ScreeningAttribute
-    overall_score: int  # 0-100
+    overall_score: float  # 0-100
     recommendation: Literal["strong_pass", "pass", "borderline", "fail"]
     summary: str
     analyzed_at: datetime
